requirements.txt
lbt-honeybee==0.2.221
rhino3dm==0.13.0
